feat: getTokId and getCtrtId
Add getTokId for CtrtID and getCtrtId for TokId. Refactor to reflect changes and fix some variable names.
